Refactor: サーバーのHookを関連するディレクトリに移動
