+++ internal/di/server/hook/http_server_hook.go
@@ -1,0 +1,77 @@
+package hook
+
+import (
+	"context"
+	"strconv"
+	"time"
+
+	"boilerplate-go/internal/config"
+	"boilerplate-go/internal/di/lifecycle"
+	"boilerplate-go/internal/di/server/extension"
+	"boilerplate-go/internal/logging"
+
+	"github.com/labstack/echo/v4"
+)
+
+const serverCallerSkip = 3
+
+// RegisterHTTPServerHooks は、HTTPサーバーを起動します。
+func RegisterHTTPServerHooks(
+	e *echo.Echo,
+	reg lifecycle.Registrar,
+	log logging.Logger,
+	appCfg *config.ApplicationConfig,
+	secCfg *config.SecurityConfig,
+	srvCfg *config.ServerConfig,
+	osCfg *config.OperationSystemConfig,
+	// 下記はサーバー機能の拡張が適用されたことを示すトークン
+	_ *extension.AppliedServerExtends,
+) {
+	reg.RegisterStart(newStartServerFunc(e, srvCfg, log, secCfg, appCfg, osCfg))
+	reg.RegisterStop(newStopServerFunc(e, log, osCfg))
+}
+
+// newStartServerFunc は、HTTPサーバーを起動する関数を生成します。
+func newStartServerFunc(
+	e *echo.Echo,
+	srvCfg *config.ServerConfig,
+	log logging.Logger,
+	secCfg *config.SecurityConfig,
+	appCfg *config.ApplicationConfig,
+	osCfg *config.OperationSystemConfig,
+) func(context.Context) error {
+	return func(_ context.Context) error {
+		addr := srvCfg.Port()
+		go func() {
+			if err := e.Start(":" + strconv.Itoa(addr)); err != nil {
+				log.Named("server.Start").Error("failed to start http server", logging.Error("e.Start", err))
+			}
+		}()
+		log.Named("server.Start").CallerSkip(serverCallerSkip).Info(
+			"http started",
+			logging.String(logging.EventTypeKey, logging.EventTypeStart),
+			logging.Time(logging.EventAtKey, time.Now()),
+			logging.String(logging.EventTzKey, osCfg.TimeZone()),
+			logging.String("port", strconv.Itoa(addr)),
+			logging.Strings("allowed_origins", secCfg.AllowedOrigins()),
+			logging.String("cidr", secCfg.CIDR().IP.String()),
+			logging.String("mode", appCfg.Mode()),
+		)
+		return nil
+	}
+}
+
+// newStopServerFunc は、HTTPサーバーを停止する関数を生成します。
+func newStopServerFunc(
+	e *echo.Echo, log logging.Logger, osCfg *config.OperationSystemConfig,
+) func(context.Context) error {
+	return func(ctx context.Context) error {
+		log.Named("server.Stop").CallerSkip(serverCallerSkip).Info(
+			"http stopping",
+			logging.String(logging.EventTypeKey, logging.EventTypeEnd),
+			logging.Time(logging.EventAtKey, time.Now()),
+			logging.String(logging.EventTzKey, osCfg.TimeZone()),
+		)
+		return e.Shutdown(ctx)
+	}
+}
